requirements/base.txt
# Core
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-dotenv==1.1.1

# Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0

# Database
psycopg2-binary==2.9.9

# File handling
Pillow==11.3.0

# Production
gunicorn==23.0.0
whitenoise==6.9.0
python-memcached==1.62

# Celery
celery==5.5.3
redis==6.4.0

# Security
dj-database-url==3.0.1
django-environ==0.12.0

# API Documentation
drf-yasg==1.21.10
coreapi==2.3.3
